pdf-parsing-agent/requirements.txt
google-adk>=1.29.0
PyMuPDF>=1.27.0
pdfplumber>=0.11.0
pytesseract>=0.3.10
Pillow>=10.0.0
pydantic>=2.0.0
pytest>=8.0.0
